+++ CHANGELOG.md
@@ -11,6 +11,11 @@
 - Add CWatM water demand to new data catalog (and remove from the old one).
 - Add `--profiling` option to `geb build/update/alter`.
 - Fix: fix for farm sizes that are all on the high end of the distribution.
+- Fix: fix for regions with very large coastal areas beyond the riverine grid
+- Use figures path for sfincs model to save all figures
+- Switch to hourly values for extreme value statistics
+- Use maximum of one flood peak per week
+- In evaluate make a dataframe without missing timesteps and ensure that return periods are esimated on the same data for observed and simulated for comparison.
 - Fix: fix for regions with very large coastal areas beyond the riverine grid.
 - Fix: waterbody outflow is larger than waterbody storage (due to floating point imprecision).